wasmparser: do not allocate when calculating lowered signatures.
This commit changes component validation in `wasmparser` so that the
calculation of a lowered function signature for validation component function
definitions doesn't allocate.

It uses a small helper type of static size to populate the list of lowered
types, bailing out early from the operation if the maximum number of parameters
or results is reached.