wasmparser: fix variant lowering calculation.
This commit fixes variant lowering when calculating lowered function
signatures.

A `get_mut` call was being made on underlying storage space for the
`LoweredTypes` helper type, which always returned success, causing the case
types to never be appended on to the original list of lowered types.

This commit fixes this by implementing `get_mut` on `LoweredTypes` and also
improves the error message for when there is a lowered signature mismatch when
validating lifting core functions.