Remove module linking support in the `wast` crate
I'm currently in the process of taking over #529 and getting it ready to
land. In an effort to make the diff smaller and easier for others to
review from this point going forward my thinking is that it would
probably be best to try to land it piecemeal if possible. To that end
the first thing I think that can be separated out would be to remove the
current module linking parsing implementation in the `wast` crate. The
design of the component model parsing support was already to delete most
of this (and re-add it back if necessary) so I don't think this will
conflict with #529 beyond "just take the #529 version". This I think
makes it a bit easier to review the diff, however.