Reorganize the `wast` crate
This commit is a complete reorganization of the `wast` crate to make
room for the component module implementation upcoming in #529. Along the
lines of the `wasmparser` crate I feel like the best way to implement
this is to have a `core` module with all the core wasm parsing along
with a `component` module which will have all the component parsing.
The shuffles happening in this commit are:

* The `wast::ast::token` module is now `wast::token`
* The `wast::ast::{wast, assert_expr, wat}` modules and types are now
  present at the root of the crate.
* All other contents of `wast::ast` are moved to a new `wast::core` module.
* The `wast::resolve` module is mostly moved into `wast::core::resolve`
  (still private from external users)
* The `wast::Error` type is now defined in an `error` submodule and
  reexported to the root.

No functional change is intended from this commit, instead this is
internally changing style (using `core::*` imports instead of
`core::Foo` namings) and updating imports to use the new locations. My
hope is that most of the component model can live in a new
`wast::component` submodule with light integration elsewhere e.g. in the
`Wat` type.

Finally I also added a CI job that runs `cargo doc` since I believe
there were a bunch of broken rustdoc links before this commit, and
that's something we should weed out on CI.