Simplify the testing harness
This commit aims to simplify the test harness as it has grown quite
complicated over the years. The original intention was to simplify
insertion of component handling into the wast test harness and it
snowballed into some other improvements first that I wanted to get out
of the way. Namely here the improvements are:

* All `*.wasm` files are removed from the test suite. They're either
  translated to `*.wat` or `*.wast` files since the `module binary`
  form can encode any binary (with comments) as well.

* The difference between `AssertMalformed` and `AssertInvalid` is
  removed and the same arm now handles both of these directives. Perhaps
  one day in the spec test suite these meant something different but
  nowadays the tests seem inconsistent enough that there's no clear
  distinction between the two (e.g. the error "must happen in this
  phase" vs another). In general I don't think precisely where an error
  pops up is too important and we already have a few cases of
  divergence, so as long as things are rejected with the right error
  message I figure it's ok.

* The `match` for wast directives is now exhaustive to ensure handling
  of any new directives that are added.

* The `WastDirective::QuoteModule` directive is merged directly into
  `WastDirective::Module`. Additionally `QuoteModule` now has a helper
  `encode` method to avoid the need to actually match on the module
  itself and delegate appropriately (e.g. the parsing is baked into the
  `wast` crate now.

* Support for `*.txt` tests are removed since this is a vestige of
  wabt-based testing.

* Support for tests with `invalid` in the name are removed now that
  they're all transitioned to `*.wast` tests and can use the normal
  testing harness.

Overall I'm hoping that this will make it easier to replace the `module`
argument of `assert_malformed` and `assert_invalid` with a `Wat`
perhaps, and similarly replacing the `Module` directive with `Wat` or
something like that too. That way the test harness here won't need any
changes to update for components since by-and-large the testing simply
takes an input, encodes it, and then tests the result or the error that
pops out.