wasm-encoder: update to the latest component model changes.
This commit updates and refactors `wasm-encoder` to include the latest
component model changes.

Most notably, the index spaces for core and component items in the component
model have been separated.

As a result, some of the component section encoders have been split in two,
with a "core" version going into the `core` module; for things like aliases and
instance sections, the core versions may one day be used in module encoding if
a standalone module linking proposal is put forth.