wasmparser: fix instance validation.
This commit fixes instance validation so that only module and instance types
exist within the instance index space.

Previously, the instance index space in the validator might have had module,
component, or instance types. This lead to confusing validation.

To fix this, a new `ModuleInstanceType` is introduced to represent an instance
of a module. Instantiation of a module now only inserts a `ModuleInstanceType`
into the instance type space. Additionally, instantiation of a component only
inserts an `InstanceType` in the instance type space.

This makes it much easier to reason with what types might be present in the
instance index space.

Fixes #586.