wasm-encoder: reduce duplicated code for encoding sections.
This commit is a refactoring of `wasm-encoder` such that an `Encode` trait is
added and implemented by every core and component section.

Most sections `Encode` implementation is delegated to a single `encode_section`
function. This reduces the amount of duplicated code.

Additionally, moving the implementation of `ComponentImportSection` out of the
core module where it doesn't belong.