Remove `ItemRef` and `IndexOrRef` from core `wast` support
These constructs are vestiges of the old module linking proposal where
any reference to any item could be replaced with
`(kind $instance "export")` as sugar for injectin aliases. At this time
core wasm does not support nested instances so this removes the support
in the `wast` crate to bring the core wasm support more in line with the
specification.