wasm-encoder: introduce an `Encode` trait implemented on all encodable types (#609)
* wasm-encoder: minor lint fixes.

* wasm-encoder: reduce duplicated code for encoding sections.

This commit is a refactoring of `wasm-encoder` such that an `Encode` trait is
added and implemented by every core and component section.

Most sections `Encode` implementation is delegated to a single `encode_section`
function. This reduces the amount of duplicated code.

Additionally, moving the implementation of `ComponentImportSection` out of the
core module where it doesn't belong.

* wasm-encoder: implement `Encode` for remaining encodable types.

This commit removes private `encode` methods in favor of implementing the
`Encode` trait for the remaining encodable types.

* wasm-encoder: change `Encode` trait to take `&mut Vec<u8>`.

* wasm-parser: restore `ModuleSection` and `NestedComponentSection`.

This commit adds these two sections back to `wasm-encoder` to make the
intention of sub-components and sub-modules explicit.

It does change their type to store a reference rather than a `Component` or
`Module`.