Refactor the `wast` component resolver (#614)
* Fix a failing test on `main`

Currently `(start $f)` fails to compile because the loop that parses
values expected a `(result ..)` to always be present, but it is not
alwyas guaranteed to be present.

* Refactor the `wast` resolver

The purpose of this commit is to cut down on the complexity of reading
and understanding the resolver for components. These were some
refactorings I noticed during review that I think help make it a bit
more clear in each function what's happenign.

* Add a new `Resolver` type to encapsulate the `Vec<ComponentResolver>`
  stack
* Move freestanding functions to methods on `Resolver` and remove
  lifetime arguments on each function.
* Rename the `ComponentResolver` type to `ComponentState` to avoid
  conflicting with `Resolver`.
* Move `register*` functions to `ComponentState`

* Refactor alias injection in component resolution

This refactors the injection of aliases to share the same code across
components themselves, component types, and instance types. This way the
drain of `aliases_to_insert` is only processed in one location to avoid
having to worry about it from multiple locations.