Bump crate versions (#616)
* Bump crate versions

To help facilitate development of the component model in Wasmtime itself
I wanted to release the current implementation of components in `*.wast`
plus recent fixes for wasmparser/wasmprinter/etc. This'll all continue
to change but publishing the current state should help unblock the
Wasmtime changes.

* Fix a failing test on `main`

Currently `(start $f)` fails to compile because the loop that parses
values expected a `(result ..)` to always be present, but it is not
alwyas guaranteed to be present.