wasmparser: update for the component model spec changes.
This commit refactors `wasmparser` to conform to the latest component model
spec.

The biggest change is that core items and component items have distinct index
spaces and sections. This had a large impact on the validator.

Various type names have been changed to better align with the terminology used
in the component model spec ("interface" is now gone in favor of component
"value types").