wast: update for the component model proposal changes.
This commit basically rewrites the component model support in wast for the
latest text format and binary format changes.

`wast` now depends on `wasm-encoder` for encoding components so that there is a
single implementation of the component binary format.

In the future, we can also replace the core encoding implementation with using
`wasm-encoder` instead (once it supports the GC proposal encoding).