wast: Add declared subtyping to type definitions
The GC proposal type system has declared subtyping where
a definition indicates the parent type(s) it is a subtype
of [1]. For now a type is restricted to just one parent type.

There is no text format for this yet, so I tried to pick a
reasonable starting point.

e.g.

(sub $parent (type (struct)))

[1] https://github.com/WebAssembly/gc/blob/main/proposals/gc/MVP.md#type-definitions