wasm-encoder: move component-related core sections.
This commit moves the component-related "core" sections out of the core module
and into the component module in `wasm-encoder`.

Originally these were put in the core module in anticipation of one day having
the sections be part of core, but that's a long way off and may not occur.
Until that time, it makes more sense to keep these sections grouped with the
other component sections.

This also replaces `ComponentAliasKind` with `ComponentOuterAliasKind`, which
should only allow types, modules, and components to be aliased according to the
current spec.

Updated `wast` based on these API changes (the other crates will have similar
changes soon).