wasm-encoder: remove unnecessary `Export` type.
This commit removes the core `Export` type, which is simply the same
enumeration as `ExportKind`, but paired with the exported item's index.

Instead, encoding takes both the kind and index rather than an `Export`.

With this change, encoding core exports functions like component exports in
`wasm-encoder`.

Also fixed where `Export` was used in the other tools.