wast: implement identifier resolution for variant cases.
This commit updates the parsing and resolution of variant cases for the
`refines` clause.

Variant cases may now have an optional identifier on them, which is referred to
in a `refines` clause, rather than having to use the variant name string.

This makes it clearer to users of the text format that a resolution is taking
place.