Implement module/component alias sugar when instantiating
The syntax for instantiation changed slightly in #621 where a bare index
is now required, but this syntax prevents the inline injection and usage
of an alias of an export of another component/instance. This commit
soups up the support here by bringing back the `IndexOrRef` structure
which can be used to parse one of the two, allowing both a bare
reference plus a parenthesized reference that enables alias injection.