wasmparser: fix validation of `realloc` option.
This commit fixes the validation of the `realloc` option for canonical
functions.

Previously all parameters and the result was checked to see if the option was
required. However, the parameters should only be checked for lifting and the
result should only be checked for lowering.

Refactored things slightly so that `ComponentFuncType.lower` returns a struct
with named fields rather than a tuple with too many unnamed values.

Fixes #629.