Fix `arbitrary_loop`'s minimum bound
Currently if the closure bails out for whatever reason we might not
reach the minimum bound required by `arbitrary_loop` when generating a
module. This commit updates this bail-out to return an error instead
since the data is not formatted correctly to allow for the minimum
number of loop iterations to get processed.