Add allowed_instructions to SwarmConfig   (#663)
* Add `allowed_instructions` to `SwarmConfig`

This commit updates `SwarmConfig` to use the input to filter what
kinds/classes of instructions can be generated. This should
theoretically help explore the instruction space a bit more by having
generated modules where entire classes of instructions are disabled or
are the only ones enabled.

* Enable some features by default in the `smith` CLI

These are all in stable wasm itself so turn them on by default.