Remove type management in `wasmprinter`
Recently I found myself trying to debug an invalid component but the way
that it was invalid meant that I couldn't print it with `wasmprinter`.
This commit removes all the type management in `wasmprinter` which is
currently needed for the `start` function of components (and other
since-removed items I believe historically) to allow printing more
components, namely those that even may be invalid. In general I
personally think it's best for `wasmprinter` to print as much as
possible, even if it's invalid, and leave validation primarily to the
parsing phase later.

This commit does effectively "break" the implementation of printing
modules with `start` functions, however. The `(result ..)` production is
never printed any more since the type information here is not tracked.
This is tracked at WebAssembly/component-model#41 for updates to the
`start` function.