Allow getting the types that have been parsed so far from the validator
Ref #271 - after looking at `WasmModuleResources`, I realised that it's heavily geared towards what's needed for validating the code section, and so doesn't include any of the component types. So instead of implementing that, I've added a separate mechanism to get types from a `Validator` midway through parsing.

I've implemented this with a `TypesRef` type, which is meant as a borrowed version of `Types`.

Unfortunately, this does mean duplicating the methods between them; I had hoped to make `Types` deref to `TypesRef`, but `Deref` requires producing a *reference* to the target type.

I was able to unify the implementation though, with a regular method to go from `&Types` to `TypesRef`.

I didn't include any of the methods for getting the counts of things on `TypesRef`, with the reasoning that the final numbers aren't known at that point; I don't know if that really makes sense though so I'm fine to add them.

In general I'm not set on any aspect of the API, this is just the first thing I came up with.