wast: fix parsing of component value type use in item sigs.
This commit fixes the parsing of value type use in a value import.

Previously, the parser parsed it as `(import "" (value 0))` instead of the
expected `(import "" (value (type 0)))`.

The fix is to split out the parsing of component value types by index such that
every other place can parse with the index, but inside of the item signature it
parses it with the surrounding type sort.