wasm-encoder: support already-encoded instructions
wasm-encoder already has a budding support for splating raw bytes into a
section of a module as raw bytes. Right now it seems quite supported on
a fairly ad-hoc basis, but is extremely useful when transformations are
being applied.

One particularly high-value addition to this support that I found is
with `Instructions` – there are a ton of these and, when modifying a
module, most of these will be copied over, with only a couple of
interesting instructions being inspected and/or modified.

Right now this requires effectively a HUGE match statement spanning
hundreds of lines to implement well. `Instruction::Raw` would provide
a reasonable escape hatch here, much like e.g. `RawSection` and such do.