Add a first pass at support for invoking components in `*.wast`
This commit updates the `*.wast` parsing in the `wast` crate with a
first pass of parsing `assert_return` for both value arguments and value
returns for component-model-based values. The syntax for now is pretty
crude and seems highly likely to change, but I wanted to get something
started to help with tests in Wasmtime.