wasmparser: fix validation for realloc option.
The fix for #684 was a little too broad, applying the requirement that the
realloc option be present when the maximum number of flat parameters is
exceeded to both lifts and lowers.

In fact, this is only true of lifting a function and not for lowering a
function because the adapter just reads the parameters out of the caller's
linear memory and doesn't need to allocate.

Fixes #693.