add VisitOperator trait
The trait method identifiers have been automatically generated using the `heck` crate.
This is the same that I would have used if I were to generate the trait using a proc. macro.
A proc. macro was not used since I wanted to keep the PR as simple and contained as possible.

I figured all those trait methods do not necessarily need docs since that would have been very daunting work to add.
Also docs for all the trait methods can be added later.

One downside of the non-generated trait is that we are going to have to keep those identifiers in sync manually.