implement VisitOperator for FuncValidator
This was relatively straight forward and most of the work could be automated.
This basically forwards all the methods to the underlying OperatorValidator impls and adds offsets to respective errors as does the original implementation.
I prevent rustfmt from doing its job here so that we can do bulk line based transformations of the code.