wasmparser: rename init_expr to offset_expr (#698)
Some of the expressions representing an offset into a memory or table
entity are named `init_expr`. This has caused me a significant amount of
confusion multiple times now. Each time I ended up having to stop
working on my code and going to re-read the spec and the implementation
of wasmparser.

Naming these expressions `offset_expr` would make it much clearer what
they represent and put them closer to both the spec wording and wat
representation (`(offset ...)`) as well.

This is a breaking change though.