Update example to something substantive.
This commit updates the composition demo to include a simple HTTP service that
uses a configurable backend to talk to.

The backend in the demo expects a `text/plain` body and echoes the body in the
response.

With a simple change to how the component gets composed, a middleware component
can be inserted between the service and the backend components; it is
responsible for gzip compressing the response from the backend.