wasm-compose: reimplement how the tool works.
This commit completely reimplements how `wasm-compose` functions.

Previously a configuration file was required for the tool to function and it
relied on a lot of explicitly specified information to function.

Now the configuration file is optional and the tool will search for
dependencies from a set of search paths.

Dependencies that cannot be located will remain as instance imports in the
composed component; thus, the composed component can now use host-provided
interfaces.

The component now always exports every export from the root component rather
than having to be told which instance to export from.