`wasmparser`: VisitOperator API (#697)
* add VisitOperator trait

The trait method identifiers have been automatically generated using the `heck` crate.
This is the same that I would have used if I were to generate the trait using a proc. macro.
A proc. macro was not used since I wanted to keep the PR as simple and contained as possible.

I figured all those trait methods do not necessarily need docs since that would have been very daunting work to add.
Also docs for all the trait methods can be added later.

One downside of the non-generated trait is that we are going to have to keep those identifiers in sync manually.

* implement BinaryReader::visit_operator method

Note: This was primarily copy pasted from the respective match construct used so far.

* implement VisitOperator for OperatorValidator

This was the biggest chunk of the work and it was also not trivial to copy over from the `match` based code constructs.
I tried to deduplicate most of the same validation logic by splitting out those parts into their own functions such as OperatorValidator::check_v128_shift_op etc.

We might want to use these VisitOperator impl in the respective `match` based impementation to deduplicate validation logic.
This refactoring should be simple since we will only forward to the respective visitor based method.
The minor downside is that we can no longer merge match arms while forwarding.

* implement VisitOperator for FuncValidator

This was relatively straight forward and most of the work could be automated.
This basically forwards all the methods to the underlying OperatorValidator impls and adds offsets to respective errors as does the original implementation.
I prevent rustfmt from doing its job here so that we can do bulk line based transformations of the code.

* add OperatorsReader::visit_with_offset method

This is the API to the new visitor based operator validation.

* apply rustfmt

* deduplicate validation logic in OperatorValidator

* remove invalid debug_assertion in OperatorValidator

* use visitor based validation in FuncValidator::validate method

* remove FuncValidator::visit auxiliary function

This is due to Alex C. mentioning that its closure based usage is pretty heavy on the Rust comiler.
Benchmarks show that performance is not affected by this change.

* remove unused import

* move WasmModuleResources into OperatorValidator

This was demanded by Alex C. during code review.
In general this cleaned up a lot of code in the OperatorValidator impl.
However, this design is incompatible with the way the ModuleState::check_init_expr method works and causes a borrow check error there.
I have not yet been able to fully resolve this borrow check error since I am missing information about the involved behaviours.

* fix borrow issues by using input pair for OperatorValidator

This PR again moves the WasmModuleResources out of the OperatorValidator but keeps many of the improvements by now taking over a pair of (offset, resources): (usize, T) for all visitation methods.

* add default implemented VisitOperator::visit_operator method

Also use it from OperatorValidator::process_operator method.

* implement BinaryReader::read_operator in terms of VisitOperator trait

This required adding a lifetime to the VisitOperator trait to preserve the lifetime on the BrTable.

* prevent auto formatting for VisitOperator::visit_operator default impl

* prevent auto formatting for VisitOperator impl for OperatorFactory

* apply rustfmt

* fix compilation with --features deterministic

* fix compilation without --feature deterministic and without

* make OperatorFactory private

* remove OperatorValidator::process_operator

Replaced by direct use of VisitOperator::visit_operator call.

* remove #[inline(always)] annotation from VisitOperator::visit_operator method

* change back to impl trait syntax for some OperatorValidator methods

* add new OperatorValidatorTemp type and conversion

* finish OperatorValidatorTemp VisitOperator impl

* remove generic Input parameter from VisitOperator

* rename input -> offset in VisitOperator trait definition

* apply rustfmt

* Simplify the operator validator

* Move `Stack` back into `OperatorValidator`
* Move most methods to `OperatorValidatorTemp` to avoid passing around
  `&impl WasmModuleResources`.
* Implement `Deref` to avoid lots of `self.inner`
* Move back to one method for push/pop control/operands

Co-authored-by: Alex Crichton <alex@alexcrichton.com>