Optimize `local` lookup in the validator
Previously `wasmparser`'s validation would avoid creating a
`Vec<ValType>` for locals to prevent to avoid allocating too much
memory, instead creating a binary-searchable-array for local indices.
Most wasm functions, however, have very few locals and frequently access
them. This commit instead creates a linear flat array for the first N
locals and when the size of locals exceeds that count it falls back to
the old binary search method.

Locally this improve the validation time for the benchmarks in #704 by
up to 10%.