Don't check value types on `push_operand`
During validation the `push_operand` operation is quite hot and
shouldn't do too much extra work. This commit removes the
`check_value_type` call during that operation, instead deferring it to
the "edges" where value types may be introduced. This commit extends the
testing for this and adds necessary `check_value_type` in the other
locations that value types can be defined within a wasm module.

Overall this gives about a 10% win on the validation benchmarks added
in #704