Fix validation of `catch`/`catch_all` with `try` params
This commit fixes an accidental regression from #697 in the refactoring
I added at the end where manual modification of `self.control` was
replaced with `self.push_ctrl`. For the exceptions proposal specifically
this is not valid since the `block_type` for the frame doesn't get the
parameters pushed when the `catch` and `catch_all` instructions are
encountered.