Fix a typo in parameter orderings
This fixes a bug when using the `Operator`-based interface to the
`Validator` where the `mem` and `segment` indices were accidentally
swapped for the `memory.init` instruction. This commit goes a bit
further though and reorders method parameters to match their order of
appearance in the binary encoding, swapping the parameters in both the
`memory.init` and `memory.copy` visit functions.