Add a feature to `wasmprinter` to print binary offsets
I often find myself trying to debug error messages here and there
related to low-level validation. These error messages have an offset
within the original file of the offending opcode, but finding that
opcode within the text format is typically not trivial. This was the
original motivation for `wasm-tools dump` but that output is also quite
verbose and not always easy to parse. Consequently this problem
motivates this commit.

This commit adds a `-p` flag to the `wasm-tools print` subcommand which
will print, in comments at the start of each line, the binary offset of
the corresponding item in the binary. This notably prints the binary
offset of all instructions and should ideally make it much easier to use
just the text format to find the offending instruction given a
particular offset.