Perform some renamings within the `wasmparser` crate (#727)
* Rename `wasmparser::MemoryImmediate` to `MemArg`

I personally find `MemoryImmediate` a bit long to type and look at and
doesn't convey much more than `MemArg`. In the interest of keeping the
`VisitOperator` trait a bit more compact in its definition as well this
renames the to a shorter `MemArg` name.

* Remove the `SIMDLaneIndex` type alias

This removes the `SIMDLaneIndex` name in favor of just using `u8` in the
respective locations. Currently it's already a type alias and I don't
think the extra-long name is buying much over just using the primitive
type.

* Rename `VisitOperator` methods to match instruction names

I noticed a few discrepancies were translated from the `Operator`
names onto the `VisitOperator` trait, so this commit does a pass and
aligns all the `visit_*` names to match the instruction names themselves
as currently specified in the text format.