Fix position report for multi-byte opcodes
This commit updates position listed for errors related to multi-byte
operators to point to the first byte of the operator rather than the
discriminant byte of the operator after the prefix. This should be a
pretty minor change but helped me to better correlate an error message
with the precise start of an opcode recently.