Remove the usage of `Operator` in const validation
This commit updates the `check_const_expr` function to remove the usage
of the `Operator` visitor in the spirit of eventually removing the
`Operator` type if possible. This was achieved through some macro-magic
and isn't necessarily more readable than the prior version, so I could
see this going either way.

In the long run with #733 it's not really difficult per-se to maintain
an `Operator` interface in `wasmparser` since it's trivially defined via
macros. This may be a case where we want to use that more than the
`visit_*` pieces perhaps.

cc #711