Remove a redundant check in `check_memory_index`
Previously this checked for indices greater than 0 to gate multi-memory
but that limit happens at memory-definition time. That means that
there's no need to have a separate check when checking indices beyond
the normal bounds check. This doesn't impact validation performance at
all but it was a minor clean-up I noticed.